% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R
\docType{class}
\name{Psiform-class}
\alias{Psiform-class}
\title{An S4 class containing predictors (x), response (y) and sample weights (w)}
\description{
An S4 class containing predictors (x), response (y) and sample weights (w)
}
\section{Slots}{

\describe{
\item{\code{c}}{Phenotype values.}

\item{\code{beta}}{Parameters for genes.}

\item{\code{u}}{U matrix.}

\item{\code{v}}{V matrix.}

\item{\code{sigma_sq}}{Variances of the residual terms.}

\item{\code{param}}{Model parameters.}
}}

