% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-2-methods.R
\name{show}
\alias{show}
\alias{show,Psiform-method}
\title{Method show for the package}
\usage{
\S4method{show}{Psiform}(object)
}
\arguments{
\item{object}{A \code{Psiform} object.}
}
\description{
The method show for \code{Psiform} object.
}
\seealso{
\code{\link{Psiform-class}}\cr
}
